# python-flask-server-generated/swagger_server/models/errors.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Errors(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, status: int=None, type: str=None, message: str=None):  # noqa: E501
        """Errors - a model defined in Swagger

        :param status: The status of this Errors.  # noqa: E501
        :type status: int
        :param type: The type of this Errors.  # noqa: E501
        :type type: str
        :param message: The message of this Errors.  # noqa: E501
        :type message: str
        """
        self.swagger_types = {
            'status': int,
            'type': str,
            'message': str
        }

        self.attribute_map = {
            'status': 'status',
            'type': 'type',
            'message': 'message'
        }
        self._status = status
        self._type = type
        self._message = message

    @classmethod
    def from_dict(cls, dikt) -> 'Errors':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Errors of this Errors.  # noqa: E501
        :rtype: Errors
        """
        return util.deserialize_model(dikt, cls)

    @property
    def status(self) -> int:
        """Gets the status of this Errors.


        :return: The status of this Errors.
        :rtype: int
        """
        return self._status

    @status.setter
    def status(self, status: int):
        """Sets the status of this Errors.


        :param status: The status of this Errors.
        :type status: int
        """

        self._status = status

    @property
    def type(self) -> str:
        """Gets the type of this Errors.


        :return: The type of this Errors.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this Errors.


        :param type: The type of this Errors.
        :type type: str
        """

        self._type = type

    @property
    def message(self) -> str:
        """Gets the message of this Errors.


        :return: The message of this Errors.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this Errors.


        :param message: The message of this Errors.
        :type message: str
        """

        self._message = message