# python-flask-server-generated/swagger_server/models/temperature_target_get.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class TemperatureTargetGet(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, value: int=None):  # noqa: E501
        """TemperatureTargetGet - a model defined in Swagger

        :param id: The id of this TemperatureTargetGet.  # noqa: E501
        :type id: str
        :param name: The name of this TemperatureTargetGet.  # noqa: E501
        :type name: str
        :param value: The value of this TemperatureTargetGet.  # noqa: E501
        :type value: int
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'value': int
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'value': 'value'
        }
        self._id = id
        self._name = name
        self._value = value

    @classmethod
    def from_dict(cls, dikt) -> 'TemperatureTargetGet':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TemperatureTargetGet of this TemperatureTargetGet.  # noqa: E501
        :rtype: TemperatureTargetGet
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this TemperatureTargetGet.

        the unique identifier for the zone  # noqa: E501

        :return: The id of this TemperatureTargetGet.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this TemperatureTargetGet.

        the unique identifier for the zone  # noqa: E501

        :param id: The id of this TemperatureTargetGet.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this TemperatureTargetGet.


        :return: The name of this TemperatureTargetGet.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this TemperatureTargetGet.


        :param name: The name of this TemperatureTargetGet.
        :type name: str
        """

        self._name = name

    @property
    def value(self) -> int:
        """Gets the value of this TemperatureTargetGet.

        the temperature in \"Kelvin\", step=0.1K  # noqa: E501

        :return: The value of this TemperatureTargetGet.
        :rtype: int
        """
        return self._value

    @value.setter
    def value(self, value: int):
        """Sets the value of this TemperatureTargetGet.

        the temperature in \"Kelvin\", step=0.1K  # noqa: E501

        :param value: The value of this TemperatureTargetGet.
        :type value: int
        """
        if value is None:
            raise ValueError("Invalid value for `value`, must not be `None`")  # noqa: E501

        self._value = value